Add pagination logic

+++ app/src/main/java/com/arctouch/codechallenge/data/service/TmdbApi.kt
@@ -21,7 +21,7 @@ interface TmdbApi {
     fun upcomingMovies(
         @Query("api_key") apiKey: String,
         @Query("language") language: String,
-        @Query("page") page: Long,
+        @Query("page") page: Int,
         @Query("region") region: String
     ): Single<UpcomingMoviesResponse>
